requirements.txt
aenum==3.1.15
anyio==4.4.0
appnope==0.1.4
APScheduler==3.10.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgihandler==0.5.9
asgiref==3.11.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
biopython==1.79
bleach==6.1.0
branca==0.7.2
Brotli==1.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.3
comm==0.2.2
concurrent-log-handler==0.9.28
constantly==15.1.0
crispy-bootstrap5==2024.10
cryptography==40.0.2
cssselect2==0.8.0
cycler==0.10.0
daphne==4.0.0
dbf==0.99.9
debugpy==1.8.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.28
django-apscheduler==0.6.0
django-braces==1.14.0
django-ckeditor==6.7.1
django-cors-headers==3.13.0
django-crispy-forms==1.9.2
django-environ==0.11.2
django-extensions==4.1
django-filter==2.4.0
django-import-export==4.3.13
django-jalali==7.4.0
django-jalali-date==2.0.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-mptt==0.18.0
django-request-logging==0.7.5
django-rest-knox==4.1.0
django-sslserver==0.22
django-widget-tweaks==1.4.8
djangorestframework==3.16.1
djangorestframework-csv==2.1.1
djangorestframework-xml==2.0.0
djangorestframework_simplejwt==5.5.1
drafthorse==2025.1.1
drf-base64==2.0
drf-extra-fields==3.1.1
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
et-xmlfile==1.1.0
exceptiongroup==1.3.1
executing==2.0.1
factur-x==3.6
Faker==8.13.2
fastjsonschema==2.20.0
filetype==1.2.0
folium==0.17.0
fonttools==4.58.1
future==0.18.2
getmac==0.8.2
GridDataFormats==0.5.0
gsd==2.4.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
humanize==4.12.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.0.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.3
isodate==0.7.2
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.17.2
Jinja2==3.1.4
joblib==1.0.1
json5==0.9.25
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
Khayyam==3.0.17
kiwisolver==1.3.2
lxml==5.4.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.7
mistune==3.0.2
mmtf-python==1.1.2
msgpack==1.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.6.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
overrides==7.7.0
packaging==26.0
pandas==1.5.1
pandocfilters==1.5.1
parso==0.7.1
pdf2image==1.17.0
persiantools==5.4.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.2.2
pluggy==1.6.0
portalocker==3.2.0
priority==1.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.7
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-machineid==0.5.1
pyarmor==7.3.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.8.0
pymodbus==3.8.6
PyMySQL==1.0.2
pyodbc==5.1.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pypdf==5.6.0
PyPDF2==1.26.0
pyphen==0.17.2
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.19.3
pyserial==3.5
pytest==8.4.2
pytest-django==4.11.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.2.1
python-json-logger==2.0.7
python-memcached==1.62
pytz==2022.5
PyYAML==6.0
pyzmq==26.0.3
qrcode==7.4.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
reportlab==3.6.6
requests==2.28.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
schedule==1.2.2
scipy==1.13.1
Send2Trash==1.8.3
service-identity==21.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.5
stack-data==0.6.3
tablib==3.9.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.4.0
tornado==6.4.1
tqdm==4.62.3
traitlets==5.0.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.15.0
tzdata==2022.5
tzlocal==2.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.12
waitress==3.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.11
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
xyzservices==2024.6.0
yapf==0.31.0
zeep==4.3.2
zipp==3.19.2
zope.interface==6.0
zopfli==0.2.3.post1
